rename color to index
